requirements.txt
streamlit
torch
ultralytics
Pillow
requests
opencv-python-headless